UPDATE-JS-HEADER
Added header;
Added JS folder;
Added JS files;
Imported JS Files;

малі фікси хедера ще дороблю
