+++ js/explore-our-new-chocolate-first.js
@@ -1,0 +1,11 @@
+const refs = {
+    openMenuBtn: document.querySelector('[data-explore-open]'),
+    closeMenuBtn: document.querySelector('[data-explore-close]'),
+    menu: document.querySelector('[data-explore]'),
+  };
+  refs.openMenuBtn.addEventListener('click', toggleMenu);
+  refs.closeMenuBtn.addEventListener('click', toggleMenu);
+  function toggleMenu() {
+    refs.menu.classList.toggle('is-hidden');
+    document.body.classList.toggle('no-scroll');
+  }
