+++ src/css/header.css
@@ -19,7 +19,7 @@
         align-items: center;
         justify-content: center;
         font-weight: 700;
-        font-size: 14px;
+        font-size: 12px;
         line-height: 1.29;
         letter-spacing: -0.03em;
         color: var(--color-black);
@@ -28,7 +28,6 @@
     
     
     .logo-first-part {
-        display: inline-block;
         font-style: italic;
     }
 
@@ -44,6 +43,9 @@
         justify-content: center;
         background-color: transparent;
         border: none;
+        transition-duration: 250ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-property: stroke;
         
     }
 
@@ -53,6 +55,9 @@
         stroke: var(--color-black);
         width: 28px;
         height: 28px;
+        transition-duration: 250ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-property: stroke;
     }
     
     .mob-menu {
@@ -60,7 +65,7 @@
         width: 59%;
         max-width: 290px;
         height: 100vh;
-        position: absolute;
+        position: fixed;
         top: 0;
         right: 0;
         display: flex;
@@ -70,11 +75,11 @@
         padding: 20px;
         align-items: center;
         justify-content: center;
+        transition-duration: 250ms;
+        transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+        transition-property: opacity;
     }
     
-    .mob-menu.menu-open {
-        display: flex;
-    }
 
 
     .mob-nav {
@@ -94,6 +99,14 @@
         line-height: 1.14;
         letter-spacing: -0.02em;
         color: var(--color-white);
+        transition-duration: 250ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-property: color;
+    }
+
+    .nav-link:hover,
+    .nav-link:focus {
+        color: var(--color-orange);
     }
     
     .mob-menu-close {
@@ -103,6 +116,9 @@
         display: flex;
         align-items: center;
         justify-content: center;
+        transition-duration: 250ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-property: stroke;
     }
     
     .menu-close-icon {
@@ -110,7 +126,19 @@
         width: 28px;
         height: 28px;
         stroke: var(--color-white);
+        transition-duration: 250ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-property: stroke;
     }
+
+    .mob-menu-close:hover .menu-close-icon,
+    .mob-menu-close:focus .menu-close-icon,
+    .mob-menu-open:hover .mob-menu-icon,
+    .mob-menu-open:focus .mob-menu-icon  {
+        stroke: var(--color-orange);
+    }
+
+    
 
     .social {
         margin-top: auto;
@@ -126,6 +154,9 @@
     .social-icons {
         stroke: var(--color-white);
         fill: transparent;
+        transition-duration: 250ms;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-property: stroke;
     }
 
     .social-link:hover .social-icons,
@@ -143,7 +174,15 @@
         height: 20px;
     }
 
+    @media screen and (min-width: 375px) {
+        .logo {
+            font-size: 14px;
+        }
+    }
+
     @media screen and (min-width: 768px) {
+
+        
 
         .social-link {
             width: 24px;
@@ -170,8 +209,8 @@
         }
 
         .header {
-            padding: 24px 0px 32px 0px;
-          }
+            padding: 28px 0px 36px 0px;
+        }
 
         .nav-link {
             color: var(--color-black);
