+++ src/partials/header.html
@@ -9,7 +9,7 @@
         </div>     
         <ul class="nav mob-tab-hidden"> 
             <li class="nav-item"><a href="../index.html" class="nav-link">Home</a></li>
-            <li class="nav-item"><a href="#" class="nav-link">How it’s made?</a></li>
+            <li class="nav-item"><a href="#" class="nav-link">How it&apos;s made?</a></li>
             <li class="nav-item"><a href="#" class="nav-link">Our products</a></li>
             <li class="nav-item"><a href="#" class="nav-link">Top sellers</a></li>
             <li class="nav-item"><a href="#" class="nav-link">Chocolate is loved</a></li>  
@@ -32,7 +32,7 @@
                 <use href="../images/icons.svg#hamburger"></use>
             </svg>
         </button>
-        <div class="mob-menu dt-hidden" data-menu>
+        <div class="mob-menu dt-hidden is-hidden no-scroll" data-menu>
             <button type="button" class="mob-menu-close" data-menu-close>
                 <svg class="menu-close-icon">
                     <use href="../images/icons.svg#close"></use>
